# More efficient reading of the next character